Add ability to read last email for testing
https://eaflood.atlassian.net/browse/CMEA-279

Currently, when signing off the TCM for release we have to manually test anything related to the email functionality.

For example, the password reset function relies on sending a link in an email. This means we have to send the email to a real address to be able to access it to follow the link.

Adding the ability to read the last email to the TCM would mean we can automate the tests instead.

This is exactly what we did with the WCR, WEX and FRAE services which are also built using Rails and Ruby. In fact, we built a tool to make it possible for any Rails service to have this feature; [defra-ruby-email](https://github.com/DEFRA/defra-ruby-email).

This change adds **defra-ruby-email** to the project so that we can access the last email sent in our acceptance tests.