Add LastEmailCache
This is a singleton class which has 2 purposes

- it is registered by `ActionMailer::Base.register_observer` (more on that in next commit)
- it has all the functionality to extract the content from the ActionMailer message hash

It's a singleton because we just want one instance of the last email sent across the app.

The code is essentially a straight copy of [Defra Ruby Email](https://github.com/DEFRA/defra-ruby-email/blob/main/lib/defra_ruby_email/last_email_cache.rb) and what it does. Unfortunately, because the project hasn't been updated inawhile when we tried to use it

- it reverted Rails and its dependencies to v6
- it errored if github_changelog_generator wasn't a dependency in the project

So, for now we're resorting to bringing in a copy of the code directly.