Add the LastEmailObserver
> Register an Observer which will be notified when mail is delivered. Either a class, string or symbol can be passed in as the Observer. If a string or symbol is passed in it will be camelized and constantized.

So, as long as whatever is passed in provides a `delivered_email()` method you're good. It's there you put whatever you need to do with the sent email. In our case that's setting the last email instance in our LastEmailCache singleton.

We found it _must_ go in the initializers folder else the observer never seemed to get registered even though the call was included in other files that are autoloaded.

Again, this is a straight copy of what is in Defra Ruby Email.