Add the last email route
The endpoint which allows us to see the content of the last email. Thing of interest is the constraint. This means rather than the endpoint always existing but we just chuck an error when running in production. To be a little more secure the route simply doesn't exist which means any malicious user would get a 404 and hopefully be put off trying to dig further.