Add ability to read last email for testing (#533)
https://eaflood.atlassian.net/browse/CMEA-279

Currently, when signing off the TCM for release we have to manually test anything related to the email functionality.

For example, the password reset function relies on sending a link in an email. This means we have to send the email to a real address to be able to access it to follow the link.

Adding the ability to read the last email to the TCM would mean we can automate the tests instead.

This is exactly what we did with the WCR, WEX and FRAE services which are also built using Rails and Ruby. In fact, we built a tool to make it possible for any Rails service to have this feature; [defra-ruby-email](https://github.com/DEFRA/defra-ruby-email).

This change was intended to add **defra-ruby-email** to the project but we encountered issues doing so.

- it reverted Rails and its dependencies to v6
- it errored if github_changelog_generator wasn't a dependency in the project

So, instead, we had to copy the code from **defra-ruby_email** and implement it directly into the project (we'll try and update and then implement **defra-ruby-email** another time).