Add the new endpoint and supporting functionality
Pretty much a copy of what we did for the import. Key difference is instead of `call()` returning the results (which in hindsight breaks the ServiceObject template) we add a new `results` attribute.

This gets populated by the service and returned by the endpoint. Sticking with the same convention as the imports we name the property in our JSON object `succeeded`.