Add build Docker image Github Action to CI
https://eaflood.atlassian.net/browse/CMEA-286

Now we are able to build a Docker image of the TCM we want to have our CI automatically build one when

- a change is made to `main`
- a new version [tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging) is pushed

On the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) we have this ability. We use GitHub actions and store the built image in [GitHub Container Registry](https://docs.github.com/en/packages/learn-github-packages/introduction-to-github-packages).

This change adds the same functionality to this project using the Charging Module as a template.