Update Airbrake project key env var
Just a quick houeskeeping change. When we started pulling together the task definition files for the TCM in preparation for deploying the TCM to AWS ECS we used the ones from the Charging Module API as a template. We spotted it used a slightly different name and decided we'd rename the env var for the TCM.

- this would save some effort in setting up the templates
- would bring some consistency across the 2 projects