Fix sonarcloud properties file (#538)
Over the last few PRs SonarCloud has been shouting at us because we keep making changes to config files and it can find no test coverage for it. This didn't follow our experience on other Rails-based projects so we did a quick comparison and found some inconsistencies in the files.

This change brings `sonar-project.properties` in line with those other projects.