Remove Coverband
5 months ago we [Add coverband support to the project](https://github.com/DEFRA/sroc-tcm-admin/pull/479). The intent was to learn what features what not being used by users. We could then look to start reducing the amount of code we are currently maintaining. The hope is this would lead to improved unit test coverage.

We ideally would have got a release out sooner but we are due to get one out on 9 March 2022. In the meantime we are working on migrating the project to a brand new AWS ECS environment. We've decided that there is no point in increasing the complexity of that work by requiring Redis to also be part of the new environments. Coverband depends on it so its still needed for our release to our EC2 environment on the 9 March. We believe between 9 March and the new ECS environment going live there will be plenty of time to generate and see the Coverband metrics, especially as we are heading into annual billing (a once a year event).

But we won't need it after that so we're taking the opportunity to scrub it from the project now.