Add test-only reset DB endpoint
https://eaflood.atlassian.net/browse/CMEA-281

The nature of the TCM is that it gets its data from import files. It's not data added through the UI by user interaction. This means we rely on fixed fixture files we drop into AWS S3 buckets that the TCM then grabs and imports.

This means the data in our test files is very static. Added to that once a file has been imported the data in it can't be imported again. So, in order to build repeatable tests that don't become super-complex we've worked on the basis that each test run starts with a _clean_ DB. The process of cleaning is currently done via `rake db:reset && rake db:seed` triggered through Jenkins. But this needs to be done manually. We really need to be able to trigger the reset from our test suite [sroc-acceptance-tests](https://github.com/DEFRA/sroc-acceptance-tests). That way we can just keep re-running tests with no manual intervention.

So, this change adds an endpoint that will clean the DB of existing data. It will only be accessible in our non-production environments and only accesible by admin users. We avoid using the `rake db:reset` as that drops the DB and recreates it and our experience has found that then forces you to deal with trying to kill active DB connections.

Instead we drop down an abstraction and use the PG gem to make direct calls to specific tables to truncate them ready for reuse.