Add our CleanDbService
We're still able to use ActiveRecord rather than go down to the PG gem to TRUNCATE our tables fortunately.

We only want to clean certain tables as others

- are required for Rails to work, for example `ar_internal_metadata` and `schema_migrations`
- we don't want to clear our static data such as regimes and users we've already seeded

The service works by getting a list of tables, iterating through them and only truncating if the table name is not in our list of tables to keep.

We keep a list of the tables we've truncated so we have something to return in the endpoint we're going to add next.