Use TCM_ENVIRONMENT for all test endpoints (#545)
[Add ability to read last email for testing](https://github.com/DEFRA/sroc-tcm-admin/pull/533) and [Add test-only reset DB endpoint](https://github.com/DEFRA/sroc-tcm-admin/pull/544) added endpoints that are only to be used to support testing in our non-production environments.

In PR #533 we used the env var `ENABLE_LAST_EMAIL` to determine if the function should be enabled i.e. we're in a non-production environment. We were all set to do the same for resetting the DB but instead went for something that explicitly tells the TCM what environment it's in.

Going forward, we can use this one env var to control all our test support rather than having an env var for each. We can also (if we find the time!) add the value to our Airbrake errors. That way we know exactly what non-production environment an error originated in.