Switch to GOV.UK Notify for email
https://eaflood.atlassian.net/browse/CMEA-291

Currently, the TCM sends emails via [SendGrid](https://sendgrid.com/) using SMTP. When running in an AWS environment it's a requirement that anything sent via SMTP must go via [postfix](https://www.postfix.org/). Postfix is a Mail Transfer Agent (MTA). It has lots of features but our understanding is that we primarily use it to monitor and log all SMTP traffic. This means that along with the measures we have in place for HTTP all traffic from the TCM and its environment is monitored. Bascially, security!

We're currently working on moving the TCM to AWS ECS, a Docker based environment. The use of SMTP for email is not a direct blocker to this. But _if_ we retain it we'd need to think about either trying to run a Postfix instance within our AWS ECS environment. Or, more likely, retaing our EC2 instance and figuring out how to connect the 2.

The issue goes away if we can send our email over HTTP. Sendgrid provide a web api and we could use that. But so does GOV.UK Notify and we have a number of other Defra services already using it. Retaining SMTP will mean investing time getting Postfix to work. Else we invest that time in updating the TCM to use a web API.

We've opted for the latter, and of our options we're going with GOV.UK Notify. This change covers the work involved.