Add a Notify email rake test task
It's always handy to be able to fire off an email from an app via the command line as a means of confirming that email is setup as necessary. We have done this in other apps. So, here we implement a rake task that calls a mailer that fires off a test email.

Assuming it reaches Notify and we can see it in the dashboard then we have confirmation everything is working as expected.