Bring invitation mailer inline with other mailers
First, all the other mailers are labelled as instructions so we do the same with the invitation. They also take or generate what they need from the object passed in. Invitation on the overhand was working all that out in the controller and then passing it through as args.

We refactor the code so that the mailer does all the work using the `User` instance passed in.