Exclude comment lines from line length limits
We generally adhere to our 120 char limit for all lines but there are times a comment will include an example or link that breaks that. We don't really want to break it up so this additional setting in our rubocop config tells it to ignore comment lines when evaluating this cop.