Simplify Devise from settings
As long as we have been dealing with Devise it has been a requirement that the DEVISE_MAILER_SENDER env var is set. Not doing so not only means emails don't send, but rake tasks then errored when loading.

All it actually comes down to is

- you _must_ specify `config.mailer_sender` in the Devise initializer
- because it's an email address feeling it needs to be configurable when actually it's the same for all environments and very, very unlikely to ever change

By just putting the email address in directly we reduce the complexity and can stop worrying about DEVISE_MAILER_SENDER not being present.

N.B. We tried just setting `config.mailer_sender` in the initilizer and deleting it from UserMailer. But then `from:` was missing from the mail messages being created. So, currently we are having to set it in both places.