Switch to GOV.UK Notify for email (#546)
https://eaflood.atlassian.net/browse/CMEA-291

Currently, the TCM sends emails via [SendGrid](https://sendgrid.com/) using SMTP. When running in an AWS environment it's a requirement that anything sent via SMTP must go via [postfix](https://www.postfix.org/). Postfix is a Mail Transfer Agent (MTA). It has lots of features but our understanding is that we primarily use it to monitor and log all SMTP traffic. This means that along with the measures we have in place for HTTP all traffic from the TCM and its environment is monitored. Basically, security!

We're currently working on moving the TCM to AWS ECS, a Docker-based environment. The use of SMTP for email is not a direct blocker to this. But _if_ we retain it we'd need to think about running a Postfix instance within our AWS ECS environment. Or, more likely, retaining our EC2 instance and figuring out how to connect the 2.

The issue goes away if we can send our email over HTTP. Sendgrid provides a web API and we could use that. But so does GOV.UK Notify and we have a number of other Defra services already using it. Retaining SMTP will mean investing time in getting Postfix to work. Else we invest that time updating the TCM to use a web API.

We've opted for the latter, and of our options, we're going with GOV.UK Notify. This change covers the work involved.

** Notes

The comments throughout the code cover the following in more detail. But to summarise

- We add our own custom delivery method to ActionMailer
- NotifyMail is our implementation of the custom mailer (note its required `deliver!()` method)
- We update the config for development and production to use this new method
- We tell Devise to use our mailer rather than its default
- In our mailer we add matching methods that Devise expects to call. These implementations support our NotifyMail delivery method rather than the standard SMTP one
- We create plain text versions of the Devise mailers. Emails will not actually use them; those templates are stored in Notify. But they allow us to generate email content that matches what we have saved in Notify
- We add a rake task that allows us to quickly test emails are working

On top of that we do some tidying up

- Update the invitation mailer and how it is called to match what we are doing in the other mailers
- Update the Last email functionality so that we can see notify properties and the response when we call `/last-email`
- We remove ApplicationMailer. Email headers are controlled by Notify so it is redundant
- Remove unused Devise mailer views
- Remove unsed Devise views

We also simplify the Devise `from:` settings. As long as we have been dealing with Devise it has been a requirement that the `DEVISE_MAILER_SENDER` env var is set. Not doing so not only means emails don't send, but rake tasks then errored when loading. All it actually comes down to is

- you _must_ specify `config.mailer_sender` in the Devise initializer
- because it's an email address feeling it needs to be configurable when actually it's the same for all environments and very, very unlikely to ever change

By just putting the email address in directly we reduce the complexity and can stop worrying about DEVISE_MAILER_SENDER not being present. N.B. We tried just setting `config.mailer_sender` in the initilizer and deleting it from UserMailer. But then `from:` was missing from the mail messages being created. So, currently we are having to set it in both places.

We tweaked our rubocop settings so that comment lines are excluded from the 120 char limit.

Finally, the `schema.rb` got updated when running migrations. This happened when creating the database locally. We're guessing someone update the defaults for precision in timestamp fields.