Seed additional test data
We're doing work on the [sroc-acceptance-tests](https://github.com/DEFRA/sroc-acceptance-tests) to make them repeatable and remove any need for human interaction. One of the blockers is that we need to be able to set the database to a known state for each test run. We've built features into the TCM to support this ([Add test-only reset DB endpoint](https://github.com/DEFRA/sroc-tcm-admin/pull/544), [Add non-UI endpoint to trigger file import process](https://github.com/DEFRA/sroc-tcm-admin/pull/467), [Add non-UI endpoint to trigger file export process](https://github.com/DEFRA/sroc-tcm-admin/pull/534)) but there are still some tests we can't run because of missing data.

One of these is the annual billing functionality. This is a feature that was built for Go live to import billing data directly into the service. It was a one time use thing that could be removed. It was agreed with the business the ability to still review the results needed to be retained but we could [Remove Annual billing upload functionality](https://github.com/DEFRA/sroc-tcm-admin/pull/416).

So, for testing we have a feature which depends in data being present to use it but no way to get it into the system. This is one example of test data we need to be able to seed the TCM with.