Fix Docker broken by missing Notify key
We didn't spot that [PR #546](https://github.com/DEFRA/sroc-tcm-admin/pull/546) broke our Docker workflow. Because we has used `ENV.fetch()` to grab the env var but then hadn't set it we were getting an exception thrown. Where we don't have a default but we are fine with `nil` we should be using `ENV[]` to grab the value. That won't error if the env var cannot be found.