Fix broken change password view (#555)
In [PR #546](https://github.com/DEFRA/sroc-tcm-admin/pull/546) as part of implementing GOV.UK Notify for emails we removed the Devise mailer views that were no longer being used. Whilst we were at it we also removed the unused Devise views. Turns out not all of those views were unused! 🤦

This change adds back in the edit registration view which we must have amended in the past to be our change password view.