Add non-UI endpoint to trigger data export process
https://eaflood.atlassian.net/browse/CMEA-299

https://eaflood.atlassian.net/browse/CMEA-287

To make testing the TCM easier we want to be able to trigger some tasks from our [sroc-acceptance-tests](https://github.com/DEFRA/sroc-acceptance-tests) that currently are done in the background using CRON.

For example, having loaded the system with test data if we can generate the data export file and upload it to the AWS S3 bucket from our tests we can then run assertions to confirm the data export is working as expected.

This change adds a new endpoint that when hit will trigger the data export process in the same way the CRON schedule does. There won't be any UI as it's expected to be used by the acceptance tests (though it also gives us the flexibility to trigger the job manually when needed!)