Fix issue with missing pids/ folder (again!) (#630)
We recently made the change [Fix issue with missing pids/ folder for puma](https://github.com/DEFRA/sroc-tcm-admin/pull/627) and thought all our issues with the `tmp/` folder and puma needing to create a pids folder were sorted. Well, that did nothing!

🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦🤦

The change we made was to the source code. We ensured the folder got committed. Then, when running locally using docker-compose all was well. We'd obviously not had enough ☕️ that day because it wasn't solving the problem. When running locally our compose file was instructing Docker to mount the project folder into the container. So, hey-presto, we have a tmp folder. When we finally clocked that's what we were doing and removed the mount section, the error appeared again.

A _proper_ check of the solution has brought to light the problem.

- our .dockerignore was excluding `tmp/` from being added to the built image
- even if we did add it because we run everything using our **app** user, it wouldn't have permission to write to the folder anyway

So, this change

- tidies up our .gitignore entries around `tmp/` and `log/`
- adds `tmp/` into our .dockerignore as something to be retained
- adds `chown` to the Dockerfile to ensure the **app** user has full access to `tmp/`