Switch to GOV.UK Notify for email
> This was originally done in the `main` branch as [PR 546](https://github.com/DEFRA/sroc-tcm-admin/pull/546). However, it is tied up with changes we made to support Docker and we are still waiting for our first TCM AWS ECS environment from web-ops. We have now reached a point that emails sent to Defra/EA via Sendgrid are failing more often than succeeding and now they are not even getting through to peoples Junk folders.
>
> The TCM team need to be able to invite new users to the service and we don't really have any viable workaround. So, we're pushing on with making these changes as a hotfix using the currently released version as a base.

https://eaflood.atlassian.net/browse/CMEA-291

Currently, the TCM sends emails via [SendGrid](https://sendgrid.com/) using SMTP. When running in an AWS environment it's a requirement that anything sent via SMTP must go via [postfix](https://www.postfix.org/). Postfix is a Mail Transfer Agent (MTA). It has lots of features but our understanding is that we primarily use it to monitor and log all SMTP traffic. This means that along with the measures we have in place for HTTP all traffic from the TCM and its environment is monitored. Basically, security!

We're currently working on moving the TCM to AWS ECS, a Docker-based environment. The use of SMTP for email is not a direct blocker to this. But _if_ we retain it we'd need to think about running a Postfix instance within our AWS ECS environment. Or, more likely, retaining our EC2 instance and figuring out how to connect the 2.

The issue goes away if we can send our email over HTTP. Sendgrid provides a web API and we could use that. But so does GOV.UK Notify and we have a number of other Defra services already using it. Retaining SMTP will mean investing time in getting Postfix to work. Else we invest that time updating the TCM to use a web API.

We've opted for the latter, and of our options, we're going with GOV.UK Notify. This change covers the work involved.