Update Gemfile to match lock
We had spotted that we were running with versions of gems that were higher than specified in the Gemfile, for example, Rails was on v7 but the Gemfile makes it look like we are running v6.

So, as part of some housekeeping we update the Gemfile to reflect what we are now locked to.