Fix asset pre-compile in Docker build (#677)
We now have an AWS environment we can deploy the production build of our Docker image. Only it immediately broke!

The error was related to a missing application.css asset which led us to look at what the `asset_builder` stage was doing in our Dockerfile.

We could see when we targeted `production` using `docker build` that the resulting image just had an empty sprockets manifest file generated in `/usr/src/app/public/assets`. But when we started our docker development environment and ran the same `bundle exec rake assets:precompile` everything worked perfectly.

We finally tracked it down and it was a proper 🤦🤦🤦🤦🤦 moment. For some reason that we can't fathom now, we'd told docker to ignore `app/assets` when copying the source code into the image! 😱

This meant `bundle exec rake assets:precompile` was running fine, only there were no assets to compile. The reason it worked when running the docker development environment is because our `docker-compose.yml` mounts the repo into the container i.e. `app/assets` exists when the command is run.

Whilst debugging we also tweaked the `COPY` command. It wasn't needed but it keeps things consistent with how we do the gems so we left the change in.