Update Dependabot config to check GitHub actions
When looking to clean up the warning messages we are getting when our GitHub actions run we spotted some were to do with the version of actions being used.

That's when we realised the TCM unlike our other repos does not have Dependabot configured to check the GitHub actions we are using.

This change fixes that.