Push built images to AWS ECR instead of GCR (#683)
The new environments we're building for the SROC TCM which will be Docker based will pull their images from AWS ECR. We have just been storing the images in GitHub Container Registry (GCR) whilst we have been waiting for it to be built.

We now have an ECR repo and login credentials. This is our first attempt to build and push our images to it.

Though the [Docker login-action](https://github.com/docker/login-action#aws-elastic-container-registry-ecr) supports AWS ECR, it only does so using a Key and a Secret Access Key. Our web-ops team and AWS prefer it if we instead assume a role using IAM which is why we are now using the [configure-aws-credentials](https://github.com/aws-actions/configure-aws-credentials) and [amazon-ecr-login](https://github.com/aws-actions/amazon-ecr-login) actions.

Thanks to this post; [Build And Push Docker Image to AWS ECR using GitHub Actions](https://medium.com/@anshita.bhasin/build-and-push-docker-image-to-aws-ecr-using-github-actions-506e9f77f7f8) we believe it should work!