Fix repo details in Docker workflow (#685)
When we updated `.github/workflows/docker.yml` to switch from pushing to GitHub Container Registry (GCR) to AWS ECR we overlooked the need to update other details (hey, we don't touch this that often!)

We were still creating an image that was based on the GCR registry and the repo name `DEFRA/sroc-tcm-admin`. We need to build our image using the correct registry and repo details else the docker push will be trying the wrong registry!