Amend docker tagging for AWS ECR
In [Push built images to AWS ECR instead of GCR](https://github.com/DEFRA/sroc-tcm-admin/pull/683) we started the changes to switch to AWS ECR.

We'd forgotten though that the AWS ECR registry that has been setup enforces tag immutability. This means we can't rely on a tagging strategy that uses `:main`, `:production` or `:latest`. There must be a unique 1-to-1 reference between images and tags.

Our tagging strategy before was

- let [metadata-action](https://github.com/docker/metadata-action) generate tags using it's default settings
- test if the action was because a new tag was pushed `startsWith(github.ref, 'refs/tags/v')`
- if it wasn't then add our own custom tag to what **metadata-action** generated
- push the image with those tags

The result was

- when a tag is pushed the only tag generated is `:v1.2.3` (for example)
- when it is just a push then 2 tags are generated; `:main` and `:v3.2.1-158-g733ce43` (for example)

Now we couldn't have **metadata-action** generating `:main`. So, we went back to the docs, this time a little more informed. What we found was

- we don't need to roll our own step checking what the ref starts with and making decisions based on it
- we can get **metadata-action** to do everything we need

Using it's config we tell it the types of tags we want created and their priority. So, now we use a simple step to generate the tag we'll use for normal pushes to `main`. Then we tell **metadata-action**

- if the workflow was triggered because of a new tag, only create a [semver](https://semver.org/) based tag using whatever tag we pushed
- for everything else only create a tag using the tag we generated

No more scary bash code on the workflow, we're using the action as intended, and we can run the workflow without AWS ECR telling us off!