Add job summary and comments to Docker workflow
We started the journey to [Push built images to AWS ECR instead of GCR](https://github.com/DEFRA/sroc-tcm-admin/pull/683) a few PR's ago. We faced a few bumps, some of our own making. But now everything appears to be working as expected.

To finish things off we add detailed comments to our workflow so others can follow what we are doing. We also (because we can 😎) generate a Job Summary in the workflow. In future we should then be able to see at a glance all the details for the image we've pushed.