Fix docker.yml build labels
We spotted an intellisense warning for our Docker workflow in the "Build and push Docker image" step.

It's not caused a problem, but we're not 100% certain labels were ever correctly being assigned to the image because of this.

We know we have the syntax right for the tags, and making this change removes the intellisense warning.