Disable latest tag handling in docker workflow
As part switching to AWS ECR we realised we needed to [Amend docker tagging for AWS ECR](https://github.com/DEFRA/sroc-tcm-admin/pull/688). This is because the ECR registry we are pushing too has been configured to use immutable tags.

We are ensuring we generate only a unique tag for pushes to `main` and no longer use a `:main` tag. But when we tested pushing a new git tag we saw that [metadata-action](https://github.com/docker/metadata-action/tree/master#latest-tag) was generating a `:latest` tag as well.

By default a GitHub tag event causes metadata-action to apply the `:latest` tag. You can disable this by configuring `flavor:` which defines global behaviour for tags in the action. This change does that.