Fix local setup of docker environment (#723)
When trying to build the image from scratch locally recently we were getting an error with `bundler`. This was because we were not specifying a version for bundler, and its latest version is now incompatible with Ruby 2.7.1 which we are running.

Amending the Dockerfile to specify the version to install resolved the problem.