Changes to get back to AWS
In early 2022 the SROC team were coming to the end of development of the [SROC Charging Module API (CHA)](https://gitub.com/DEFRA/sroc-chargig-module-api). It was developed for use by another team that were not yet ready to deploy their side of the project to production.

So, the SROC dev team, looking for ways to continue improving things for the business opted to move the TCM from being EC2 based (deployed to a server) to being ECS based (deployed as a Docker image).

The team had developed the CHA to be docker-based, and all new services in Defra were expected to be built on Docker. This would allow the service to take advantage of a cheaper, future proof deployment model and environment. Being an existing service that was internal only it was also selected as a pilot for web-ops to help develop a 'standardised' ECS setup and infrastructure.

In [Add Docker support to the project](#515) the breaking change was made. From that point onwards the project would expect to be built as an image and run as a container. It could no longer be deployed to EC2 using the existing methods.

Fast forward to May 2023 and that ideal was still to be realised. The SROC delivery team were no more, having moved onto other projects. When they could they were still pitching in with the pilot. But that had grown in scope and involved many more parties so nothing tangible was ready for use. Then crisis happened. The version of the TCM deployed to `production` was using [SendGrid](https://sendgrid.com/) to handle email sending. But failures in users receiving them, or losing them in junk (not SendGrid's fault by the way, we spotted too late the original developers had never properly configured SendGrid and the AWS DNS) meant it needed sorting.

It just so happens that one of the other improvements the team made before leaving the project was to switch it from SendGrid to [GOV.UK Notify](https://www.notifications.service.gov.uk/). But the switch was implemented _after_ the breaking change to Docker. So, we branched off the [version (3.2.1)](https://github.com/DEFRA/sroc-tcm-admin/releases/tag/v3.2.1) deployed to production, replicated the change, and deployed that 'hotfix' branch.

That was where the TCM was left with those of us keeping the lights on still hopeful to migration to ECS would happen. We now have to accept that will never happen for 2 reasons

- The TCM needs an **Annual billing upload** feature that was [stripped as unneeded](#416) re-implemented and working in time to be tested for April 2024 to support changes in EPR
- Plans are already in place to move 2 of the 3 regimes TCM supports off the system in Q1 and Q2 of this year. Then work will begin on the last one

The team made a number of improvements to the TCM in the last days of the project. But they are all locked behind the breaking change.

These changes are focused on getting the TCM deployed and running using the existing Capistrano scripts and AWS EC2 infrastructure. We've also created a local dev environment which will be used for working on the TCM going forward which will necessitate some changes.